+++ .github/workflows/ci.yml
@@ -40,7 +40,6 @@ jobs:
           - ci_step: "test-fuzz"
             command: make test-fuzz
             linter-cache: false
-            wasm-cache: true
             codecov: false
           - ci_step: "integration tests coreum-modules"
             command: make integration-tests-modules
