+++ integration-tests/modules/dex_test.go
@@ -32,6 +32,7 @@ import (
 	"github.com/CoreumFoundation/coreum/v5/testutil/integration"
 	assetfttypes "github.com/CoreumFoundation/coreum/v5/x/asset/ft/types"
 	customparamstypes "github.com/CoreumFoundation/coreum/v5/x/customparams/types"
+	testcontracts "github.com/CoreumFoundation/coreum/v5/x/dex/keeper/test-contracts"
 	dextypes "github.com/CoreumFoundation/coreum/v5/x/dex/types"
 )
 
@@ -2518,6 +2519,201 @@ func TestLimitOrdersMatchingWithAssetBurning(t *testing.T) {
 	// 100k is burnt 100k remains
 	requireT.Equal(sdkmath.NewInt(100_000).String(), balanceRes.Balance.String())
 	requireT.Equal(sdkmath.NewInt(0).String(), balanceRes.LockedInDEX.String())
+}
+
+// TestReentrancyBug tests to make sure that reentrancy bug does not exist in DEX. It might happen
+// because the control is given to smart contract in middle of an order placement.
+func TestReentrancyBug(t *testing.T) {
+	t.Parallel()
+
+	ctx, chain := integrationtests.NewCoreumTestingContext(t)
+
+	requireT := require.New(t)
+
+	assetFTClint := assetfttypes.NewQueryClient(chain.ClientContext)
+	dexClient := dextypes.NewQueryClient(chain.ClientContext)
+
+	dexParamsRes, err := dexClient.Params(ctx, &dextypes.QueryParamsRequest{})
+	requireT.NoError(err)
+	dexReserver := dexParamsRes.Params.OrderReserve
+
+	admin := chain.GenAccount()
+	acc1 := chain.GenAccount()
+	acc2 := chain.GenAccount()
+
+	chain.FundAccountWithOptions(ctx, t, admin, integration.BalancesOptions{
+		Amount: chain.QueryAssetFTParams(ctx, t).IssueFee.Amount.
+			AddRaw(1_000_000_000_000),
+	})
+	chain.FundAccountWithOptions(ctx, t, acc1, integration.BalancesOptions{
+		// message + order reserve
+		Amount: sdkmath.NewInt(500_000_000_000).
+			Add(dexReserver.Amount),
+	})
+	chain.FundAccountWithOptions(ctx, t, acc2, integration.BalancesOptions{
+		Amount: sdkmath.NewInt(500_000_000_000).
+			AddRaw(100_000_000_000).
+			Add(dexReserver.Amount), // message  + balance to place an order + order reserve
+	})
+
+	codeID, err1 := chain.Wasm.DeployWASMContract(
+		ctx, chain.TxFactory().WithSimulateAndExecute(true), admin, testcontracts.DexReentrancyPocWasm,
+	)
+	requireT.NoError(err1)
+
+	// issue tokenA
+	issueMsgA := &assetfttypes.MsgIssue{
+		Issuer:        admin.String(),
+		Symbol:        "TKNA",
+		Subunit:       "ua",
+		Precision:     6,
+		Description:   "TKNA Description",
+		InitialAmount: sdkmath.NewInt(1_000_000_000_000_000),
+		Features:      []assetfttypes.Feature{},
+		URI:           "https://my-class-meta.invalid/1",
+		URIHash:       "content-hash",
+	}
+	denomA := assetfttypes.BuildDenom(issueMsgA.Subunit, admin)
+	_, err = client.BroadcastTx(
+		ctx,
+		chain.ClientContext.WithFromAddress(admin),
+		chain.TxFactory().WithGas(chain.GasLimitByMsgs(issueMsgA)),
+		issueMsgA,
+	)
+	requireT.NoError(err)
+
+	//nolint:tagliatelle // these will be exposed to rust and must be snake case.
+	issuanceMsg := struct {
+		ExtraData string `json:"extra_data"`
+	}{
+		ExtraData: denomA,
+	}
+
+	issuanceMsgBytes, err := json.Marshal(issuanceMsg)
+	requireT.NoError(err)
+
+	attachedFund := chain.NewCoin(sdkmath.NewInt(1_000_000_000))
+	issueMsgB := &assetfttypes.MsgIssue{
+		Issuer:        admin.String(),
+		Symbol:        "TKNB",
+		Subunit:       "ub",
+		Precision:     6,
+		Description:   "TKNB Description",
+		InitialAmount: sdkmath.NewInt(1_000_000_000_000_000),
+		Features: []assetfttypes.Feature{
+			assetfttypes.Feature_extension,
+		},
+		URI:     "https://my-class-meta.invalid/1",
+		URIHash: "content-hash",
+		ExtensionSettings: &assetfttypes.ExtensionIssueSettings{
+			CodeId:      codeID,
+			Funds:       sdk.NewCoins(attachedFund),
+			Label:       "testing-reentrancy-bug",
+			IssuanceMsg: issuanceMsgBytes,
+		},
+	}
+
+	denomB := assetfttypes.BuildDenom(issueMsgB.Subunit, admin)
+	_, err = client.BroadcastTx(
+		ctx,
+		chain.ClientContext.WithFromAddress(admin),
+		chain.TxFactoryAuto(),
+		issueMsgB,
+	)
+	requireT.NoError(err)
+	// get extension contract addr
+	denomBTokenRes, err := assetFTClint.Token(ctx, &assetfttypes.QueryTokenRequest{
+		Denom: denomB,
+	})
+	requireT.NoError(err)
+	tokenBExtensionAddress := denomBTokenRes.Token.ExtensionCWAddress
+
+	// send from admin to cw to place an order
+	_, err = sendFromAdmin(ctx, chain, admin, tokenBExtensionAddress, "udevcore", sdkmath.NewInt(1_000_000_000))
+	requireT.NoError(err)
+	_, err = sendFromAdmin(ctx, chain, admin, tokenBExtensionAddress, denomA, sdkmath.NewInt(100_000_000_000))
+	requireT.NoError(err)
+	_, err = sendFromAdmin(ctx, chain, admin, tokenBExtensionAddress, denomB, sdkmath.NewInt(100_000_000_000))
+	requireT.NoError(err)
+	// send from admin to acc1 to place an order
+	_, err = sendFromAdmin(ctx, chain, admin, acc1.String(), denomA, sdkmath.NewInt(100_000_000_000))
+	requireT.NoError(err)
+	_, err = sendFromAdmin(ctx, chain, admin, acc1.String(), denomB, sdkmath.NewInt(100_000_000_000))
+	requireT.NoError(err)
+
+	// send from admin to acc2 to place an order
+	_, err = sendFromAdmin(ctx, chain, admin, acc2.String(), denomA, sdkmath.NewInt(100_000_000_000))
+	requireT.NoError(err)
+	_, err = sendFromAdmin(ctx, chain, admin, acc2.String(), denomB, sdkmath.NewInt(100_000_000_000))
+	requireT.NoError(err)
+
+	// place 3 SELL order from acc1(SELL A(without extension) BUY B(with extension))
+	_, err = placeOrder(ctx, chain, acc1, dextypes.SIDE_SELL, "id1", denomA, denomB, "1", sdkmath.NewInt(120_000_000))
+	requireT.NoError(err)
+	_, err = placeOrder(ctx, chain, acc1, dextypes.SIDE_SELL, "id2", denomA, denomB, "1", sdkmath.NewInt(120_000_000))
+	requireT.NoError(err)
+	_, err = placeOrder(ctx, chain, acc1, dextypes.SIDE_SELL, "id3", denomA, denomB, "1e2", sdkmath.NewInt(99_000_000))
+	requireT.NoError(err)
+
+	// place BUY order from acc2(BUY A(without extension) SELL B(with extension))
+	_, err = placeOrder(ctx, chain, acc2, dextypes.SIDE_BUY, "hackid0", denomA, denomB, "1e1", sdkmath.NewInt(100_000_000))
+	requireT.NoError(err)
+
+	acc1ABalanceRes, err := assetFTClint.Balance(ctx, &assetfttypes.QueryBalanceRequest{
+		Account: acc1.String(),
+		Denom:   denomA,
+	})
+	requireT.NoError(err)
+	requireT.Equal("99000000", acc1ABalanceRes.LockedInDEX.String())
+}
+
+func placeOrder(
+	ctx context.Context,
+	chain integration.CoreumChain,
+	acc sdk.AccAddress,
+	side dextypes.Side,
+	id, denomA, denomB, price string,
+	amount sdkmath.Int,
+) (*sdk.TxResponse, error) {
+	placeBuyOrderMsg := &dextypes.MsgPlaceOrder{
+		Sender:      acc.String(),
+		Type:        dextypes.ORDER_TYPE_LIMIT,
+		ID:          id,
+		BaseDenom:   denomA,
+		QuoteDenom:  denomB,
+		Price:       lo.ToPtr(dextypes.MustNewPriceFromString(price)),
+		Quantity:    amount,
+		Side:        side,
+		TimeInForce: dextypes.TIME_IN_FORCE_GTC,
+	}
+	return client.BroadcastTx(
+		ctx,
+		chain.ClientContext.WithFromAddress(acc),
+		chain.TxFactoryAuto(),
+		placeBuyOrderMsg,
+	)
+}
+
+func sendFromAdmin(
+	ctx context.Context,
+	chain integration.CoreumChain,
+	admin sdk.AccAddress,
+	to string,
+	denom string,
+	amount sdkmath.Int,
+) (*sdk.TxResponse, error) {
+	sendMsg := &banktypes.MsgSend{
+		FromAddress: admin.String(),
+		ToAddress:   to,
+		Amount:      sdk.NewCoins(sdk.NewCoin(denom, amount)),
+	}
+	resp, err := client.BroadcastTx(
+		ctx,
+		chain.ClientContext.WithFromAddress(admin),
+		chain.TxFactoryAuto(),
+		sendMsg,
+	)
+	return resp, err
 }
 
 // TestTradeByAdmin tests trades by admin without limits like frozen amount.
