+++ x/asset/ft/keeper/test-contracts/contract.go
@@ -8,8 +8,6 @@ import (
 var (
 	//go:embed asset-extension/artifacts/asset_extension.wasm
 	AssetExtensionWasm []byte
-	//go:embed certik-poc/artifacts/certik_poc.wasm
-	CertikPocWasm []byte
 )
 
 // Check contract.rs for constant values defined in contract.
