+++ x/dex/keeper/test-contracts/.gitignore
@@ -1,0 +1,2 @@
+target
+artifacts
