+++ x/dex/keeper/test-contracts/contract.go
@@ -1,0 +1,11 @@
+package testcontracts
+
+import (
+	_ "embed"
+)
+
+// Built artifacts of dex.
+var (
+	//go:embed dex-reentrancy-poc/artifacts/dex_reentrancy_poc.wasm
+	DexReentrancyPocWasm []byte
+)
