+++ x/dex/keeper/test-contracts/dex-reentrancy-poc/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "certik-poc"
+name = "dex-reentrancy-poc"
 version = "0.1.0"
 authors = ["Coreum"]
 edition = "2021"
 
 exclude = [
-  "certik-poc.wasm",
+  "dex-reentrancy-poc.wasm",
   "checksums.txt",
 ]
