+++ x/dex/keeper/test-contracts/dex-reentrancy-poc/src/error.rs
@@ -1,0 +1,11 @@
+use cosmwasm_std::StdError;
+use thiserror::Error;
+
+#[derive(Error, Debug)]
+pub enum ContractError {
+    #[error("{0}")]
+    Std(#[from] StdError),
+
+    #[error("Invalid amount.")]
+    InvalidAmountError {},
+}
